Added Starter Menu UI (Check description for more detailed explanation))
Implemented Menu UI for a visual of how it might look or set up. Also used as a testing ground to see if the audio manager can still persist as it needs to remain within the hierarchy to function. Also implemented a BGM name display feature that shows the name of the track being played. It only  shows when the settings button is selected (If no one likes this feature, it will be removed). I am unable to access Tri's scripts so I couldn't make any progress there and will need to figure out how to get it to work. Will work on Gameplay HUD elements next.